fix(ci): close queue lease source acceptance
