+++ scripts/measure-dev-required-latency.mjs
@@ -16,6 +16,10 @@ import {
 } from '@kungfu-tech/buildchain-alpha/candidate-timeline';
 
 import {
+  queueAdmissionRequiredContexts,
+  validateDevRequiredLatencyBaseline,
+} from './cancel-dequeued-merge-group-runs.mjs';
+import {
   collectLatestMergedPullWindow,
   latencyOnlyEvidence,
   parseDevRequiredLatencyArgs,
@@ -34,7 +38,6 @@ const QUEUE_ADMISSION_CONTRACT_PATH = path.join(
   ROOT,
   'docs/qualification/gates/dev-queue-admission.contract.json',
 );
-
 function repositoryFromOrigin() {
   const result = spawnSync('git', ['remote', 'get-url', 'origin'], {
     cwd: ROOT,
@@ -471,46 +474,6 @@ export function summarizeMergeQueueDelivery(samples) {
             : 'merge queue delivery sample meets target',
     },
   };
-}
-
-export function validateBaseline(
-  baseline,
-  requiredContexts,
-  allowedContextAdditions = [],
-) {
-  if (baseline.$schema !== 'kungfu.dev-required-latency-baseline/v1') {
-    throw new Error('unsupported dev required latency baseline schema');
-  }
-  const expected = [...baseline.requiredContexts].sort();
-  const actual = [...requiredContexts].sort();
-  const allowed = [...new Set(allowedContextAdditions)].sort();
-  const additions = actual.filter((context) => !expected.includes(context));
-  const removals = expected.filter((context) => !actual.includes(context));
-  if (
-    removals.length ||
-    additions.some((context) => !allowed.includes(context))
-  ) {
-    throw new Error(
-      `live required contexts drifted: expected ${expected.join(', ')}, got ${actual.join(', ')}`,
-    );
-  }
-  return true;
-}
-
-export function queueAdmissionRequiredContexts(contract) {
-  if (contract?.schema !== 'kungfu.dev-queue-admission/v1') {
-    throw new Error('unsupported dev queue admission contract schema');
-  }
-  if (contract.rulesetActivation?.required !== true) {
-    throw new Error(
-      'dev queue admission contract must require ruleset activation',
-    );
-  }
-  const context = String(contract.requiredContext || '');
-  if (!/^[A-Za-z0-9][A-Za-z0-9 ._/-]{0,99}$/u.test(context)) {
-    throw new Error('dev queue admission required context is invalid');
-  }
-  return [context];
 }
 
 export function requiredContextsFromEffectiveRules(rules) {
@@ -2029,7 +1992,7 @@ async function main() {
   if (!requiredContexts.length)
     throw new Error(`no required contexts on ${options.branch}`);
   if (!options.pulls.length) {
-    validateBaseline(
+    validateDevRequiredLatencyBaseline(
       JSON.parse(fs.readFileSync(BASELINE_PATH, 'utf8')),
       requiredContexts,
       queueAdmissionRequiredContexts(
