+++ scripts/measure-dev-required-latency.test.mjs
@@ -19,7 +19,6 @@ import {
   mergeQueueEvidence,
   nativeEvidenceFromMembers,
   nearestRank,
-  queueAdmissionRequiredContexts,
   report,
   requiredContextsFromEffectiveRules,
   requiredMergeQueueWindow,
@@ -28,7 +27,6 @@ import {
   summarize,
   summarizeMergeQueueDelivery,
   summarizeNativeAttribution,
-  validateBaseline,
 } from './measure-dev-required-latency.mjs';
 import { partitionAffectedNativePlan } from './run-core-affected-native.mjs';
 
@@ -1099,46 +1097,6 @@ test('context admission ignores successful post-merge reruns', () => {
   assert.equal(context.durationMs, 300000);
   assert.deepEqual(context.workflowRunIds, [42]);
   assert.equal(context.endAuthority, 'first-success-no-later-than-pull-merge');
-});
-
-test('live required contexts match baseline or one declared lease expansion', () => {
-  const baseline = {
-    $schema: 'kungfu.dev-required-latency-baseline/v1',
-    requiredContexts: ['a', 'b'],
-  };
-  assert.equal(validateBaseline(baseline, ['b', 'a']), true);
-  assert.equal(
-    validateBaseline(baseline, ['b', 'queue-lease', 'a'], ['queue-lease']),
-    true,
-  );
-  assert.throws(
-    () => validateBaseline(baseline, ['a', 'c']),
-    /live required contexts drifted/,
-  );
-  assert.throws(
-    () => validateBaseline(baseline, ['a', 'queue-lease'], ['queue-lease']),
-    /live required contexts drifted/,
-  );
-});
-
-test('queue admission contract authorizes only one explicit required context', () => {
-  assert.deepEqual(
-    queueAdmissionRequiredContexts({
-      schema: 'kungfu.dev-queue-admission/v1',
-      requiredContext: 'Queue admission lease',
-      rulesetActivation: { required: true },
-    }),
-    ['Queue admission lease'],
-  );
-  assert.throws(
-    () =>
-      queueAdmissionRequiredContexts({
-        schema: 'kungfu.dev-queue-admission/v1',
-        requiredContext: 'Queue admission lease',
-        rulesetActivation: { required: false },
-      }),
-    /must require ruleset activation/,
-  );
 });
 
 test('required contexts come from all effective required-status rules', () => {
